Video information had been added.
- DB structure had been added
for video , comment info in "Users"
for creator info in "Videos"
for creator info in "Comments"

- Change Password page was implemented.
- Video Detail page had been added to Videos of Specific User.
We can confirm videos of user in Users' detail page.
- Video Detail page had been added to creator info.
- When User upload new video, "User" document is added to video info(Object of videos._id).

ETC...

+++ controllers/videoController.js
@@ -44,8 +44,11 @@ export const postUpload = async (req, res) => {
   const newVideo = await Video.create({
     fileUrl: path,
     title,
-    description
+    description,
+    creator: req.user.id
   });
+  req.user.videos.push(newVideo.id);
+  req.user.save();
   console.log(newVideo);
   //res.render("upload", { pageTitle: "Upload" });
   res.redirect(routes.videoDetail(newVideo.id));
@@ -57,7 +60,8 @@ export const videoDetail = async (req, res) => {
     params: { id }
   } = req;
   try {
-    const video = await Video.findById(id);
+    const video = await Video.findById(id).populate("creator"); // populate of property can be used only "Object" type.
+    // console.log(video);
     res.render("videoDetail", { pageTitle: video.title, video });
   } catch (error) {
     console.log(error);
@@ -70,7 +74,11 @@ export const getEditVideo = async (req, res) => {
   } = req;
   try {
     const video = await Video.findById(id);
-    res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+    if (String(video.creator) !== req.user.id) {
+      throw Error();
+    } else {
+      res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+    }
   } catch (error) {
     console.log(error);
     res.redirect(routes.home);
@@ -97,7 +105,12 @@ export const deleteVideo = async (req, res) => {
     params: { id }
   } = req;
   try {
-    await Video.findOneAndRemove(id);
+    const video = await Video.findById(id);
+    if (video.creator !== req.user.id) {
+      throw Error();
+    } else {
+      await Video.findOneAndRemove(id);
+    }
   } catch (error) {
     console.log(error);
   } finally {
